♻️ extract setupViewTest.specHelper.ts
